# backend/openapi/uniflow-api.yaml
openapi: 3.0.3
info:
  title: Uniflow API
  version: 1.0.0
  description: >
    Specification OpenAPI des endpoints actuellement implementes dans le backend Uniflow.
servers:
  - url: http://localhost:8080
    description: Serveur local Spring Boot
...
paths:
...
  /api/users/all:
    get:
      tags: [Users]
      summary: Lister les utilisateurs pour l'administration
      operationId: getAllUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des utilisateurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminUserDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Recuperer un utilisateur par identifiant
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Utilisateur trouve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      summary: Supprimer un utilisateur
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Utilisateur supprime
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/users/{id}/organization:
    put:
      tags: [Users]
      summary: Modifier l'organisation rattachee a un utilisateur
      operationId: updateUserOrganization
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserOrganizationUpdateRequest'
      responses:
        '200':
          description: Organisation utilisateur modifiee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/poles/create:
    post:
      tags: [Poles]
      summary: Creer un pole
      operationId: createPole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoleDTO'
      responses:
        '201':
          description: Pole cree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoleDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/poles/{id}:
    get:
      tags: [Poles]
      summary: Recuperer un pole par identifiant
      operationId: getPoleById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Pole trouve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoleDTO'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Poles]
      summary: Modifier un pole
      operationId: updatePole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoleDTO'
      responses:
        '200':
          description: Pole modifie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoleDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Poles]
      summary: Supprimer un pole
      operationId: deletePole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Pole supprime
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Donnees invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentification requise ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Acces interdit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflit avec une ressource existante
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nom:
          type: string
          example: Dupont
        prenom:
          type: string
          example: Alice
        email:
          type: string
          format: email
          example: alice@univ-lyon1.fr
        dateInscription:
          type: string
          format: date-time
          example: 2026-03-25T14:00:00
    AdminUserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Alice Dupont
        email:
          type: string
          format: email
          example: alice@univ-lyon1.fr
        roles:
          type: array
          items:
            type: string
          example: [Utilisateur, Organisateur]
        organization:
          type: string
          example: Club Informatique
        organizations:
          type: array
          items:
            type: string
          example: [Club Informatique, BDE]
        registrationDate:
          type: string
          format: date
          example: 2026-03-25
    UserOrganizationUpdateRequest:
      type: object
      properties:
        organization:
          type: string
          description: Ancien format compatible. Nom de l'organisation cible. Vide pour retirer le role organisateur.
          example: Club Informatique
        organizations:
          type: array
          description: Nouveau format. Liste des organisations cibles. Liste vide pour retirer toutes les organisations.
          items:
            type: string
          example: [Club Informatique, BDE]
...
    PoleDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          example: Informatique
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Operation effectuee avec succes
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Identifiants invalides.